20210124  意见修改

+++ ruoyi/src/main/java/com/ruoyi/project/system/service/impl/SysUserServiceImpl.java
@@ -287,14 +287,18 @@ public class SysUserServiceImpl implements ISysUserService
         user.setUpdateTime(DateUtils.getNowDate());
 
         Long userId = user.getUserId();
-        // 删除用户与角色关联
-        userRoleMapper.deleteUserRoleByUserId(userId);
-        // 新增用户与角色管理
-        insertUserRole(user);
-        // 删除用户与岗位关联
-        userPostMapper.deleteUserPostByUserId(userId);
-        // 新增用户与岗位管理
-        insertUserPost(user);
+        if(StringUtils.isNotNull(user.getRoleIds())){
+            // 删除用户与角色关联
+            userRoleMapper.deleteUserRoleByUserId(userId);
+            // 新增用户与角色管理
+            insertUserRole(user);
+        }
+        if(StringUtils.isNotNull(user.getPostIds())){
+            // 删除用户与岗位关联
+            userPostMapper.deleteUserPostByUserId(userId);
+            // 新增用户与岗位管理
+            insertUserPost(user);
+        }
         return userMapper.updateUser(user);
     }
